+++ riddle-renderer-wgpu/src/sprite_init_args.rs
@@ -1,10 +1,10 @@
-use crate::{wgpu_ext::*, *};
+use crate::wgpu_ext::*;
 
 /////////////////////////////////////////////////////////////////////////////
 // struct SpriteBuilder
 /////////////////////////////////////////////////////////////////////////////
 
-/// Builder to construct new [`Sprite`]s from `riddle_image::Image`s.
+/// Represents the settings which are used when constructing a sprite.
 ///
 /// # Example
 ///
@@ -16,22 +16,19 @@ use crate::{wgpu_ext::*, *};
 /// let renderer = Renderer::new_from_window(&window)?;
 ///
 /// let img = Image::new(100, 100);
-/// let sprite = SpriteBuilder::new(img)
-///     .with_filter_modes(FilterMode::Linear, FilterMode::Linear)
-///     .build(&renderer)?;
+/// let sprite = Sprite::new_from_image(&renderer, &img, &SpriteInitArgs::new()
+///     .with_filter_modes(FilterMode::Linear, FilterMode::Linear))?;
 /// # Ok(()) }
 /// ```
-pub struct SpriteBuilder {
-	pub(crate) img: image::Image,
-	pub(crate) mag_filter: FilterMode,
-	pub(crate) min_filter: FilterMode,
+pub struct SpriteInitArgs {
+	pub mag_filter: FilterMode,
+	pub min_filter: FilterMode,
 }
 
-impl SpriteBuilder {
-	/// Create a new builder for the given image
-	pub fn new(img: image::Image) -> Self {
+impl SpriteInitArgs {
+	/// Create a new default init args.
+	pub fn new() -> Self {
 		Self {
-			img,
 			mag_filter: Default::default(),
 			min_filter: Default::default(),
 		}
@@ -42,13 +39,5 @@ impl SpriteBuilder {
 		self.mag_filter = mag_filter;
 		self.min_filter = min_filter;
 		self
-	}
-
-	/// Build the sprite for the given renderer
-	pub fn build<Device: WGPUDevice>(
-		self,
-		renderer: &WGPURenderer<Device>,
-	) -> Result<WGPUSprite<Device>> {
-		WGPUSprite::new_from_image(renderer, &self.img, self.mag_filter, self.min_filter)
 	}
 }